Possible 0x0017/0x001C Authentication failure/Invalid sequence number fix
Thanks to Reto /  info@makerspace-reinach.ch for sharing this fix!
Unfortunately I don't have a meter to test this on anymore, but as several people are struggeling with this issue, I have updated the software. Hopefully it will fix this issue for everyone!
As I cannot test this: USE AT YOUR OWN RISK!
Best Regards
gert@dabbler.dk